chore(server): shrink duplicated HTTP and provider test harnesses
Drop overlapping app/auth smoke and share session, vendor, and tool-loader fixtures so the same route and provider behavior is asserted once.
