+++ apps/server/src/services/agent-service.test.ts
@@ -474,43 +474,6 @@ describe("AgentService vision settings", () => {
       vision: { model: "p-openai-1::gpt-4o-mini" },
     });
   });
-
-  test("does not reset coding-agent passthrough when vision is saved", async () => {
-    const db = createInMemoryDatabaseAdapter();
-    await db.upsertWorkspaceSettings({
-      codingAgentHarnesses: [],
-      codingAgentProviderPassthrough: false,
-      id: "workspace-settings",
-      imageModel: null,
-      selectedCodingAgentHarness: null,
-      transcriptionModel: null,
-      updatedAt: new Date().toISOString(),
-      visionModel: null,
-    });
-    const service = new AgentService(
-      {
-        defaultProviderId: "p-openai-1",
-        providers: [
-          {
-            apiKey: "test-key",
-            createdAt: new Date().toISOString(),
-            id: "p-openai-1",
-            label: "OpenAI",
-            type: "openai",
-          },
-        ],
-      },
-      null,
-      db
-    );
-
-    await service.setVisionSettings({ model: "p-openai-1::gpt-4o-mini" });
-
-    expect(await db.getWorkspaceSettings()).toMatchObject({
-      codingAgentProviderPassthrough: false,
-      visionModel: "p-openai-1::gpt-4o-mini",
-    });
-  });
 });
 
 describe("AgentService transcription settings", () => {
@@ -548,44 +511,58 @@ describe("AgentService transcription settings", () => {
     });
   });
 
-  test("does not reset coding-agent passthrough when transcription is saved", async () => {
-    const db = createInMemoryDatabaseAdapter();
-    await db.upsertWorkspaceSettings({
-      codingAgentHarnesses: [],
-      codingAgentProviderPassthrough: false,
-      id: "workspace-settings",
-      imageModel: null,
-      selectedCodingAgentHarness: null,
-      transcriptionModel: null,
-      updatedAt: new Date().toISOString(),
-      visionModel: null,
-    });
-    const service = new AgentService(
-      {
-        defaultProviderId: "p-openai-1",
-        providers: [
-          {
-            apiKey: "test-key",
-            createdAt: new Date().toISOString(),
-            id: "p-openai-1",
-            label: "OpenAI",
-            type: "openai",
-          },
-        ],
-      },
-      null,
-      db
-    );
-
-    await service.setTranscriptionSettings({
+  test.each([
+    {
+      field: "visionModel",
+      model: "p-openai-1::gpt-4o-mini",
+      save: (service: AgentService) =>
+        service.setVisionSettings({ model: "p-openai-1::gpt-4o-mini" }),
+    },
+    {
+      field: "transcriptionModel",
       model: "p-openai-1::whisper-1",
-    });
+      save: (service: AgentService) =>
+        service.setTranscriptionSettings({ model: "p-openai-1::whisper-1" }),
+    },
+  ] as const)(
+    "does not reset coding-agent passthrough when $field is saved",
+    async ({ field, model, save }) => {
+      const db = createInMemoryDatabaseAdapter();
+      await db.upsertWorkspaceSettings({
+        codingAgentHarnesses: [],
+        codingAgentProviderPassthrough: false,
+        id: "workspace-settings",
+        imageModel: null,
+        selectedCodingAgentHarness: null,
+        transcriptionModel: null,
+        updatedAt: new Date().toISOString(),
+        visionModel: null,
+      });
+      const service = new AgentService(
+        {
+          defaultProviderId: "p-openai-1",
+          providers: [
+            {
+              apiKey: "test-key",
+              createdAt: new Date().toISOString(),
+              id: "p-openai-1",
+              label: "OpenAI",
+              type: "openai",
+            },
+          ],
+        },
+        null,
+        db
+      );
 
-    expect(await db.getWorkspaceSettings()).toMatchObject({
-      codingAgentProviderPassthrough: false,
-      transcriptionModel: "p-openai-1::whisper-1",
-    });
-  });
+      await save(service);
+
+      expect(await db.getWorkspaceSettings()).toMatchObject({
+        codingAgentProviderPassthrough: false,
+        [field]: model,
+      });
+    }
+  );
 });
 
 describe("AgentService coding delegation context", () => {
